requirements.txt
python-dotenv
mysql-connector-python
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib